+++ test/it/shellutils/free_test.sh
@@ -1,0 +1,3 @@
+TestFreeHeader() {
+    free | head -n 1
+}