+++ test/it/shellutils/killall_test.sh
@@ -1,0 +1,6 @@
+TestKillall() {
+    sleep 30 &
+    sleep 0.2
+    killall sleep
+    echo "killed:$?"
+}