+++ test/it/shellutils/nohup_test.sh
@@ -1,0 +1,3 @@
+TestNohupRuns() {
+    nohup echo hello
+}